aprimorando a exclusão de cadastro

+++ lotofacil/app/Http/Controllers/Profiles/DeleteController.php
@@ -6,16 +6,25 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\User;
+use Illuminate\Support\Facades\Hash;
 
 class DeleteController extends Controller
 {
-    public function destroy($id) {
-       
-        if(!$user = User::find($id)) 
-        return redirect()->route('index')->with('success', 'Perfil excluído com sucesso.');
+    public function deleteAccount() {
+        $user = Auth::user();
+        return view('profiles.delete', compact('user'));
+    }
 
-        $user->delete();
-        
-        return view('index', compact('user'));
+    public function destroy(Request $request, $id) {
+        $user = Auth::user();
+
+        if (Hash::check($request->current_password, $user->password)) {
+            $user->delete();
+            Auth::logout();
+
+            return redirect()->route('index')->with('success', 'Conta excluída com sucesso!');
+        } else {
+            return redirect()->route('profile.show')->with('error', 'Senha incorreta. A conta não foi excluída.');
+        }
     }
 }
